Fix transaction get endpoints
The group by introduced in 7648fe4 had an issue when a transaction
contains duplicate items. For example, if someone pays for 2 identical coffees,
there will be two rows in transaction_item with equal fields, except for
the id column. Since the id column was not part of the group key, those
two rows would be grouped together. The get_transactions endpoint thus
only returned one item in the transaction in this case.

The fix uses grouping by primary key whereever possible.
This also allows us to remove group by on the other columns from the
same table, since PostgreSQL is smart enough to see that these must also
be unique when grouping by a primary key.